adding to db

+++ src/db/mysql/insert/pair.go
@@ -6,7 +6,7 @@ import (
 )
 
 
-func AddPairToDB(PrimaryTokenID int64, SecondaryTokenID int, PairNetworkId int, PairDexId int64, PairName string, PairAddress string) int64 {
+func AddPairToDB(PrimaryTokenID int64, SecondaryTokenID int64, PairNetworkId int, PairDexId int64, PairName string, PairAddress string) int64 {
 
 	DBKeys := "primary_token_id, secondary_token_id, network_id, dex_id, name, address"
 	SelectStatement := fmt.Sprintf("(SELECT %d AS primary_token_id, %d AS secondary_token_id, %d AS network_id, %d AS dex_id, '%v' AS name, '%v' AS address)", PrimaryTokenID, SecondaryTokenID, PairNetworkId, PairDexId, PairName, PairAddress)
